requirements.txt
slack-bolt>=1.17.0
slack-sdk>=3.19.5
python-dotenv>=1.0.0
openai>=1.3.0
requests>=2.28.2
aiohttp>=3.8.4
langchain>=0.0.267
langchain-openai>=0.0.2
fastapi>=0.100.0
uvicorn>=0.23.2